window-actor: Use int for opacity

+++ src/compositor/meta-shadow-factory-private.h
@@ -46,7 +46,7 @@ void        meta_shadow_paint       (MetaShadow            *shadow,
                                      int                    window_y,
                                      int                    window_width,
                                      int                    window_height,
-                                     guint8                 opacity,
+                                     int                    opacity,
                                      cairo_region_t        *clip,
                                      gboolean               clip_strictly);
 void        meta_shadow_get_bounds  (MetaShadow            *shadow,
